Basic Async Test setup
Moved the asynctask to it's own class and added a basic test to verify
that a string comes back from the task.
