+++ MyApplication/app/build.gradle
@@ -26,12 +26,19 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
+
+    androidTestCompile 'com.android.support:support-annotations:25.1.1'
+    androidTestCompile 'com.android.support.test:runner:0.5'
+    androidTestCompile 'com.android.support.test:rules:0.5'
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+
     compile project(':jokeactivitylib')
     compile project(path: ':jokesapi', configuration: 'android-endpoints')
-
+    compile ('com.google.api-client:google-api-client-android:1.22.0') {
+        exclude module: 'httpclient'
+    }
     compile 'com.android.support:appcompat-v7:25.1.1'
     compile 'com.google.android.gms:play-services-ads:10.0.1'
     compile 'com.android.support:support-v4:25.1.1'
