+++ MyApplication/app/src/main/java/com/example/jessemitchell/builditbigger/MainActivity.java
@@ -1,28 +1,13 @@
 package com.example.jessemitchell.builditbigger;
 
-import android.content.Context;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.v4.util.Pair;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Toast;
 
-import com.example.jesse.mitchell.builditbigger.jokesapi.myApi.MyApi;
 import com.example.jokeactivitylib.JokeActivity;
-import com.google.api.client.extensions.android.http.AndroidHttp;
-import com.google.api.client.extensions.android.json.AndroidJsonFactory;
-import com.google.api.client.googleapis.services.AbstractGoogleClientRequest;
-import com.google.api.client.googleapis.services.GoogleClientRequestInitializer;
-
-import java.io.IOException;
-
-import static com.example.jessemitchell.builditbigger.BuildConfig.APP_ENGINE_END_POINT;
-import static com.example.jessemitchell.builditbigger.BuildConfig.APP_ENGINE_PORT;
-import static com.example.jessemitchell.builditbigger.BuildConfig.APP_ENGINE_URL;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -58,7 +43,7 @@ public class MainActivity extends AppCompatActivity {
     public void tellJoke(View view) {
 
         try {
-            String joke = new EndpointsAsyncTask().execute(new Pair<Context, String>(this, "Manfred")).get();
+            String joke = new EndpointAsyncTask().execute("Manfred").get();
             // use AsyncTask to pull joke
             Intent jalIntent = new Intent(this, JokeActivity.class);
             jalIntent.putExtra("Joke", joke);
@@ -70,39 +55,4 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    class EndpointsAsyncTask extends AsyncTask<Pair<Context, String>, Void, String> {
-        private MyApi myApiService = null;
-        private Context context;
-        private String APP_ENGINE_ROOT = APP_ENGINE_URL + ":" + APP_ENGINE_PORT + APP_ENGINE_END_POINT;
-        @Override
-        protected String doInBackground(Pair<Context, String>... params) {
-            if(myApiService == null) {  // Only do this once
-                MyApi.Builder builder = new MyApi.Builder(AndroidHttp.newCompatibleTransport(),
-                        new AndroidJsonFactory(), null)
-                        .setRootUrl(APP_ENGINE_ROOT)
-                        .setGoogleClientRequestInitializer(new GoogleClientRequestInitializer() {
-                            @Override
-                            public void initialize(AbstractGoogleClientRequest<?> abstractGoogleClientRequest) throws IOException {
-                                abstractGoogleClientRequest.setDisableGZipContent(true);
-                            }
-                        });
-
-                myApiService = builder.build();
-            }
-
-            context = params[0].first;
-            String name = params[0].second;
-
-            try {
-                return myApiService.sayHi(name).execute().getData();
-            } catch (IOException e) {
-                return e.getMessage();
-            }
-        }
-
-        @Override
-        protected void onPostExecute(String result) {
-            Toast.makeText(context, result, Toast.LENGTH_LONG).show();
-        }
-    }
 }
